requirements.txt
asgiref==3.11.0
black==25.11.0
click==8.3.1
Django==5.2.8
djangorestframework==3.16.1
dotenv==0.9.9
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==12.0.0
platformdirs==4.5.0
psycopg2-binary==2.9.11
python-dotenv==1.2.1
pytokens==0.3.0
sqlparse==0.5.4
